backend/requirements.txt
fastapi
uvicorn
pydantic
motor
sqlalchemy
psycopg2-binary
redis
pika
python-multipart
pymongo
